refactored user service
