+++ springapp/src/main/java/com/examly/springapp/controller/AdmissionController.java
@@ -4,7 +4,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import com.examly.springapp.models.*;
 import java.util.*;
-import com.examly.springapp.service.*;
 import com.examly.springapp.service.admin.AdminService;
 
 @RestController
