+++ springapp/src/main/java/com/examly/springapp/controller/ReviewController.java
@@ -1,7 +1,6 @@
 package com.examly.springapp.controller;
 
 import org.springframework.web.bind.annotation.*;
-import com.examly.springapp.service.*;
 import com.examly.springapp.service.review.ReviewService;
 import com.examly.springapp.models.*;
